UPSTREAM:DP:wifi:ath12k: Add sanity check for fragment length
In backpressure scenarios, the hardware logs packets with a length
less than 16 bytes (which includes the 802.3 header and padding).
This causes an incorrect calculation when subtracting 16 from the
packet length, resulting in a large invalid value and potentially
leading to a crash.

Adding a sanity check to ensure that the packet length is greater
than 16.

Change-Id: I374dd908245c440e0583d5e5dc442559b18b59e2
Signed-off-by: Sushant Butta <sbutta@qti.qualcomm.com>