UPSTREAM:PLATFORM:wifi:ath12k: Remove SOC from group only when recovery is enabled
In case of Firmware crash when recovery is not enabled the RDDM dumps
for all the SOC's have to be collected for which ag->num_started is
used to know how many SOC's are started so that upon collecting those
many dumps host will issue a BUG_ON().

But currently in reset_work the SOC is detached from the group and
num_probed is decremented, because of which BUG_ON will be invoked
without collecting all SOC's dump.

Fix this by not calling ath12k_core_to_group_ref_put() when recovery
is not enabled so that the core won't be detached from group and the
rddm dump for all SOC's can be copied.

Change-Id: Ie1ed24f9127e2a09da0dffa38571bc6e114c2acf
Signed-off-by: Avula Sri Charan <asrichar@qti.qualcomm.com>