UPSTREAM:PROTOCOL: wifi: ath12k: Handle first AFC expiry event
During device bootup, as part of the initial WMI transactions, the
firmware sends the regulatory WMI event 'WMI_REG_CHAN_LIST_CC_EXT_EVENTID'
containing the regulatory rules for the default country. For devices that
supports AFC, this is followed by the AFC expiry event.

Currently, AFC expiry handling and AFC request generation are managed per
radio (tied to struct ath12k). However, at the time the first AFC expiry
event is received, the radio-level object is not yet allocated, causing
the event to be ignored.

This leads to unnecessary delays in AFC spectrum queries and results in
sub-optimal performance when operating in LPI power mode. To address this,
cache the AFC expiry event sent by the firmware during bootup and use it
to generate the AFC request during mac registartion. Forward this request
to higher layers to initiate querying of the AFC server.

Fixes: Ib095b796fd5 ("UPSTREAM: PROTOCOL: Add AFC Expiry event changes")
Patch-work: None
Tag: 6G_REGULATORY
Feature: AFC
Patch-dependency:
ath12k/0373-a-UPSTREAM-PROTOCOL-Define-global-6GHz-operating-class.patch
ath12k/0373-b-UPSTREAM-PROTOCOL-Prepare-complete-AFC-request-for-A.patch
ath12k/0373-c-UPSTREAM-PROTOCOL-Add-support-to-send-AFC-request-as.patch
ath12k/0373-d-UPSTREAM-PROTOCOL-Add-APIs-to-construct-AFC-frequenc.patch
ath12k/0373-e-UPSTREAM-PROTOCOL-Add-functions-to-manage-6GHz-opera.patch
ath12k/0373-f-UPSTREAM-PROTOCOL-Add-support-for-validating-and-fil.patch
ath12k/0373-g-UPSTREAM-PROTOCOL-Add-API-to-construct-AFC-location-.patch
ath12k/0373-h-UPSTREAM-PROTOCOL-Add-WMI-changes-to-handle-AFC-Expi.patch

Change-Id: I791d5d3a6adcc1bd15f1b8165e647ab6feb021d8
Signed-off-by: Thirusenthil Kumaran J <thirusen@qti.qualcomm.com>