QSDK:CP:mac80211.sh: Populate monitor interface name
With AP+Monitor+ station or mesh supplicant will try bring down
only the AP interface but not monitor interface when it's going
for a scan because of which a warning is seen in case of DFS
channels.

To bring down the monitor interface ucode needs the monitor interface
name on each radio. Added change to populate the monitor interfaces
name and pass it wpa_supplicant.uc.

Change-Id: Ic88e422003064eda5f9dce93b880265e0d3215c7
Signed-off-by: Avula Sri Charan <asrichar@qti.qualcomm.com>