UPSTREAM:CP: wifi:ath12k: Ensure proper locking when fetching arvif from ahvif list
The arvif is retrieved from the ahvif list, which is protected by RCU.
However, access to this list also requires holding the wiphy lock, as
it is used during link addition and deletion. This change ensures the wiphy
lock is acquired when accessing the ahvif list to fetch the correct arvif.

Change-Id: I8523d9aedc5f0ac8f133a6a71de93b2af1e4998f
Signed-off-by: Gautham Kumar Senthilkumaran <gauthamk@qti.qualcomm.com>