UPSTREAM:CP: wifi: mac80211: Fix the AP-VLAN update when netdev is reused
Currently whenever an AP-VLAN netdev is reused we are trying to use the
sta->valid_links bmap to fetch the mac address from vlansdata's
vif->link_conf[link_id] and update it in wdev->link[link_id].bssid
but in some cases where the previous APVLAN interfaces sdata's
vif.valid_links is not same as current STA associated we will end
up accessing link that dosent exist.

To fix this update the links first so that AP-VLAN links are updated as per
the STA valid_links and later copy the mac address to avoid such mismatch.

Change-Id: Ibd0601334a926d8e06dc769c898f32b3622e0e6e
Signed-off-by: Amith A <amitajit@qti.qualcomm.com>