UPSTREAM:CP: wifi: ath12k: Reduce power consumption during Wi-Fi down
Currently, when Wi-Fi goes down, power consumption remains high even if
no radio is active. To address this, the driver sends a WMI pdev_suspend
command to the firmware when the radio is inactive.

In the WSI bypass scenario, the SoC is bypassed if it's inactive, and the
pdev_suspend command is already sent in such cases. Remove the redundant
logic that attempts to send pdev_suspend again in the WSI bypass path.

Patch-dependency:
0441-03-UPSTREAM-CP-wifi-ath12k-Enable-pdev-suspend-and-resu.patch
Patch-work: None

Change-Id: Id9b90ad5bde692a626aa1c0c2c3401a26247bd65
Signed-off-by: Aaradhana Sahu <aaradhana.sahu@oss.qualcomm.com>