UPSTREAM:PLATFORM:wifi: ath12k: Skip core reset in case  of FW crash in rmmod path
Currently for PCI based chips core_reset is being queued in case of a
Firmware crash in rmmod path because of which ending up accessing ah
after it is freed which is leading to a NULL crash.

Also unregister the panic notifier for all the SOC's that are part of a
group
to avoid rddm dump collection after mhi suspend in case of any host panics.

Set the CRASH_FLUSH for each soc upon receiving the FW crash notification
irrespective of recovery is enabled or not and add changes to check on that
flag and collect rddm dump before mhi_stop & pci_sw_reset.

Disable SSR recovery on rmmod so that all FW crash notifications will
be receieved instead of adding additional handling for recovery enabled
cases This will also ensure that once rmmod is triggered only rootpd fatal
notification is receieved.

Remove the constraint to set the recovery mode only when radio is up so
that the SSR recovery can be disabled in rmmod case irrespective of radios
are enabled or disabled.

Change-Id: I17ed32b9e3b3e692ddff4b939b265558782a38e4
Signed-off-by: Avula Sri Charan <asrichar@qti.qualcomm.com>