QSDK:PLATFORM:wifi:ath12k: Update minidump unregister call
Adding changes for the below,

- Update minidump unregister call from the beginning of the
core_deinit call to the end of it.
This is to avoid unregistering minidump related segments
before the rmmod completed.

- Adding the fw_recovery_support flag to avoid invoking the
minidump trigger when recovery is disabled.
If recovery is disabled the minidump won't be collected when
a target assert occurs, it will be collect from U-boot
only when the dump_minimal U-boot env variable is set.

- Adding changes to fix the compilation issue when MEMDEBUG
flag is enabled.

Change-Id: I4c1892b3fbdf1a5c9d440e3af5e00046c03654a5
Signed-off-by: Raghavendran Madhavan <raghmadh@qti.qualcomm.com>