QSDK:CP: mac80211.sh: Update channel list with latest frequency range
If user changes the country code, the frequency range can also change in
iw phyXX info. This can result in failing ML reconfig as the current
script will fail to find the band with updated frequency range.

To resolve this update the channel list is wireless config using the
latest frequency range.

Change-Id: I1eed945adfab30d96d55e561f5edb88b33b97fd2
Signed-off-by: Rajat Soni <rajson@qti.qualcomm.com>