UPSTREAM:DP:wifi:ath12k: Avoid setting ath12k_dp_peer to NULL within ath12k_dp_link_peer
struct ath12k_dp_link_peer has a backpointer "dp_peer" which points to
the parent object ath12k_dp_peer. This backpointer to set to NULL in
the API ath12k_dp_link_peer_unassign during the delete of link_peer.

There can be cases wherein ath12k_dp_link_peer is found in the linked
list after unassignment and dp_peer is accessed which results in NULL
pointer access and device goes for a crash.

In order to handle this, avoid resetting of dp_peer within
ath12k_dp_link_peer in the unassignment path as ath12k_dp_peer is the
last object to be deleted in the peer delete flow.

Patch-dependency: 0116-0002-wifi-ath12k-Attach-and-detach-ath12k_dp_link_peer-to.patch

Change-Id: Ib36793d541e386e2b2f4300a3cebfa9746c60e5b
Signed-off-by: Harsh Kumar Bijlani <hbijlani@qti.qualcomm.com>