UPSTREAM:PROTOCOL: wifi:nl80211: Fix disabling of FILS/UBPR
Currently, when FILS max interval or UBPR interval is set to zero to
disable FILS or UBPR, the feature is not disabled. This is because
the parsing logic fails to indicate that there is an update for the
feature.

Fix the issue by setting the update flag to true, when the FILS
max interval or UBPR interval is zero.

Fixes: I9c352f9d42c (mac80211: Add support to enable/disable FILS/UBPR frames)

Patch-depdendency: 352-mac80211-Add-support-to-enable-disable-FILS-UBPR-fra.patch

Patch-work: None

Change-Id: I8532fc6438087d594882bd696a660806cd0882bb
Signed-off-by: Suraj P Kizhakkethil <suraj.kizhakkethil@oss.qualcomm.com>