QSDK:PROTOCOL: wifi-scripts: hostapd.sh: Set default country code to US
Currently, ieee80211d is enabled in hostapd config only if the
country code is set in UCI. As a result, the hostapd does not
advertise the country code in beacons and probe responses. This
causes connectivity issues with few mobile client devices when
operating in the 6 GHz band.

To fix this and enable ieee80211d by default, set the default
country code to US in the hostapd conf if it is not set in UCI.
If the country code is set in UCI, it will override the default
value.

US is chosen as the default country code since it is the default
country code configured in the QCA ath12k driver.

Change-Id: I175758d7bef2c7bb48912c66962235520dda4550
Signed-off-by: Thirusenthil Kumaran J <thirusen@qti.qualcomm.com>