UPSTREAM:CP: wifi:ath12k: Fix destroying the rhashtable for stale entries
Currently, arsta is maintained in a rhashtable, which is used to fetch
the entry using the station address as the key. During Mode 0 recovery,
the host sends disassociation frames to all peers connected to the AP.
After sending disassoc, it removes entries from the rhashtable one by
one using the station list from mac80211. In one instance, arsta was
freed but not removed from the rhashtable.

This causes a stale entry to remain in the table, eventually leading to
corruption of the rhashtable and a crash. Such stale entries are
dangerous and must be handled properly to maintain the integrity of the
data structure and ensure system stability during recovery operations.

To fix this, if a stale entry is found in the rhashtable, the entire
table is destroyed and reinitialized to store fresh arsta entries.
This prevents corruption, clears stale entries, and ensures the
rhashtable remains in a consistent and stable state.

Patch-dependency:
0241-0002-wifi-ath12k-WDS-subsystem-recovery-fix.patch

Change-Id: I3cf4a103c0c806d6100fd9d2dcdd6ee70a2d7b78
Signed-off-by: Gautham Kumar Senthilkumaran <gauthamk@qti.qualcomm.com>