UPSTREAM:CP: wifi: ath12k: Fix vdev start freq params for 160 MHz
The WMI command 'WMI_VDEV_START_REQUEST_CMDID' includes
frequency and width information for the channel on which the vdev is
started. There are three fields for frequency: 'mhz', 'band_center_freq1'
and 'band_center_freq2'. The 'mhz' field indicates the primary channel
frequency, while the 'band_center_freq1' and 'band_center_freq2' fields
vary based on the channel width.

For 160 MHz channel, 'band_center_freq1' should be set to the center
frequency of the primary 80 MHz segment in which the primary channel
resides, and 'band_center_freq2' should be set to the center frequency
of the entire 160 MHz channel.

But currently, 'band_center_freq1' and 'band_center_freq2' are
filled wrongly for 160 MHz channel width. 'band_center_freq1' is set to
the center frequency of the entire 160 MHz channel, and
'band_center_freq2' is set to 0.

To fix this, update the code to correctly set 'band_center_freq1'
and 'band_center_freq2' for 160 MHz channel width.

Fixes: Idd214a7c81e ("pkg upgrade Added subsys, ath11k and ath12k patches for win.wlan_host_opensource.4.0")
Patch-work: None
Patch-dependency:
ath12k/0066-wifi-ath12k-Pass-correct-values-of-center-freq1-and-.patch

Change-Id: Id8088bc92c0c682c7633fdf92eb6100088d28567
Signed-off-by: Thirusenthil Kumaran J <thirusen@qti.qualcomm.com>