QSDK:CP: wifi-scripts: mac80211.sh: Set tx power support for MLD
Currently, configuration of txpower using UCI is not supported for MLD
interfaces.

mac80211.sh script uses 'iw' tool to set the txpower for non-MLD interfaces
using the command 'iw dev <if_name> set txpower fixed <value in mBm>'.
For MLD interfaces, the iw tool shall also include the link id for which
the txpower needs to be set. The command format is as below:
'iw dev <if_name> set txpower -l <link_id> fixed <value in mBm>'.

Add support to set txpower for MLD interfaces in mac80211.sh script.

Add support to determine the link id of the MLD interface for a
configured radio and use the link id to set the txpower for the MLD
interface.

Change-Id: I397f7ae540865a92b5f418f76e6dd5de42162112
Signed-off-by: Thirusenthil Kumaran J <thirusen@qti.qualcomm.com>