QSDK:PROTOCOL:hostapd: Enable ignore_old_scan_res to flush the old scan result
Set ignore_old_scan_res=1 in the supplicant config to flush the old
scan result.

When we are not flushing the old scan results, the STA is
trying to authenticate using the older scan reults and
in the scenario when after giving wifi to AP the MAC addresses
are changing. The Auth request and response MLD addr are not
matching and it is getting added to ignore list for 3 mins.
Resulting in longer association time than expected.

Change-Id: I028661572948f1417dd3b16aba175f2cb17956ad
CRs-fixed: 4281209
Signed-off-by: Diya Sati <dsati@qti.qualcomm.com>