QSDK:CP:wifi:ath12k: Skip assigning affinity twice for the shared CE interrupt
While going through the CE's from 0 we would hit into a position
where the interrupt for one CE is reused(CE0,CE7). And assigning
affinity twice for the same irq num would alter the interrupt
handling cores for the radio. To avoid that, skip the duplicate affinity
setting for the irq

Patch-dependency: none


Change-Id: I65b37f0c58cdcb4c204f56dc386f180d430e058d
Signed-off-by: Tamizh Chelvam Raja <tamizhr@qti.qualcomm.com>